requirements.txt
pyyaml>=6.0
pydantic>=1.10.0
networkx>=2.8
jinja2>=3.1.0
click>=8.0.0
rich>=12.0.0
